requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
fpdf2>=2.7.0
tiktoken>=0.7.0
